backend/requirements.txt
fastapi==0.115.6
uvicorn==0.22.0
google-cloud-documentai==3.1.0
google-cloud-storage==2.19.0
google-auth==2.37.0
pydantic==2.10.5
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.0.1
pytesseract==0.3.10
python-multipart==0.0.20
firebase-admin==6.6.0